server/requirements.txt
anyio==3.6.1
black==22.8.0
click==8.1.3
dnspython==2.2.1
fastapi==0.85.0
flake8==5.0.4
h11==0.13.0
idna==3.4
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
pymongo==4.2.0
python-dotenv==0.21.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
